% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heliaphen_plant.r
\name{area_predict}
\alias{area_predict}
\title{Predict plant leaf area from input features extracted with image analysis.}
\usage{
area_predict(experiment, index, model, min = 0.5)
}
\arguments{
\item{experiment}{experiment code (character string)}

\item{index}{experimental design (dataframe)}

\item{model}{fitted model used for prediction (glm model object)}

\item{min}{minimal growth threshold used to discard plant from analysis (numeric, default to 0.5, i.e. +50\% growth)}
}
\value{
a dataframe for predicted plant leaf area.
}
\description{
Predict plant leaf area from input features extracted with image analysis.
}
